requirements.txt
rich
requests
python-dotenv
steam
cryptography
discord
igdb-api-v4
apscheduler
playwright